+++ gld/admin.py
@@ -1,7 +1,40 @@
 from django.contrib import admin
 
-from .models import GLD, MeasurementTvp, Observation
+from . import models as gld_models
 
-admin.site.register(GLD)
-admin.site.register(Observation)
-admin.site.register(MeasurementTvp)
+
+class GLDAdmin(admin.ModelAdmin):
+    model = gld_models.GLD
+    list_display = (
+        "bro_id",
+        "data_owner",
+    )
+
+    list_filter = ("data_owner",)
+
+
+class ObservationAdmin(admin.ModelAdmin):
+    model = gld_models.Observation
+    list_display = (
+        "gld",
+        "observation_type",
+        "begin_position",
+        "data_owner",
+    )
+
+    list_filter = ("data_owner",)
+
+
+class MeasurementTvpAdmin(admin.ModelAdmin):
+    model = gld_models.MeasurementTvp
+    list_display = (
+        "observation",
+        "data_owner",
+    )
+
+    list_filter = ("data_owner",)
+
+
+admin.site.register(gld_models.GLD, GLDAdmin)
+admin.site.register(gld_models.Observation, ObservationAdmin)
+admin.site.register(gld_models.MeasurementTvp, MeasurementTvpAdmin)